Add `_fun_keep_unit_test.py`
